--- src/main/java/tn/esprit/spring/controler/MessageController.java
@@ -1,5 +1,0 @@
-package tn.esprit.spring.controler;
-
-public class MessageController {
-	
-}
